+++ src/main/java/com/nhnacademy/book/order/controller/OrderCancellationController.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.book.order.controller.command;
+package com.nhnacademy.book.order.controller;
 
 import com.nhnacademy.book.order.dto.OrderCancelRequestDto;
 import com.nhnacademy.book.order.service.command.OrderCancellationService;
