+++ src/main/java/com/nhnacademy/book/order/controller/OrderReturningController.java
@@ -1,9 +1,12 @@
-package com.nhnacademy.book.order.controller.command;
+package com.nhnacademy.book.order.controller;
 
+import com.nhnacademy.book.order.dto.OrderReturnDto;
 import com.nhnacademy.book.order.dto.OrderReturnRequestDto;
 import com.nhnacademy.book.order.service.command.OrderReturningService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -13,6 +16,18 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 public class OrderReturningController {
     private final OrderReturningService orderReturningService;
+
+    @GetMapping("/api/orders/order-returns")
+    public ResponseEntity<Page<OrderReturnDto>> getOrderReturns(Pageable pageable) {
+        Page<OrderReturnDto> orderReturnDtoPage = orderReturningService.getAllOrderReturns(pageable);
+        return ResponseEntity.status(HttpStatus.OK).body(orderReturnDtoPage);
+    }
+
+    @GetMapping("/api/orders/order-returns/{tracking-number}")
+    public ResponseEntity<OrderReturnDto> getOrderReturn(@PathVariable("tracking-number") String trackingNumber) {
+        OrderReturnDto orderReturnDto = orderReturningService.getOrderReturnByTrackingNumber(trackingNumber);
+        return ResponseEntity.status(HttpStatus.OK).body(orderReturnDto);
+    }
